Actualizacion de campos - Ambrosio
