+++ registerSpringBoot/src/main/java/com/proyecto/encuentranos/controladores/ClienteControlador.java
@@ -1,5 +1,6 @@
 package com.proyecto.encuentranos.controladores;
 
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -11,7 +12,6 @@ import com.proyecto.encuentranos.modelos.ClienteModelo;
 import com.proyecto.encuentranos.modelos.ProveedorModelo;
 import com.proyecto.encuentranos.servicios.ClienteServicio;
 
-@CrossOrigin(origins = "http://localhost:3000/")
 @RestController
 @RequestMapping("/cliente")
 public class ClienteControlador {
