+++ registerSpringBoot/src/main/java/com/proyecto/encuentranos/controladores/ProveedorControlador.java
@@ -1,10 +1,10 @@
 package com.proyecto.encuentranos.controladores;
 
+
 import java.util.ArrayList;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,8 +15,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.proyecto.encuentranos.modelos.ProveedorModelo;
 import com.proyecto.encuentranos.servicios.ProveedorServicio;
-
-@CrossOrigin(origins = "http://localhost:3000/")
 @RestController
 @RequestMapping("/proveedor")
 public class ProveedorControlador {
