+++ registerSpringBoot/src/main/java/com/proyecto/encuentranos/controladores/ServicioControlador.java
@@ -9,7 +9,6 @@ import org.springframework.web.bind.annotation.*;
 import com.proyecto.encuentranos.modelos.ServicioModelo;
 import com.proyecto.encuentranos.servicios.ServicioServicio;
 
-@CrossOrigin(origins = "http://localhost:3000/")
 @RestController
 @RequestMapping("/servicio")
 public class ServicioControlador {
