+++ registerSpringBoot/src/main/java/com/proyecto/encuentranos/controladores/ServicioPrestadorControlador.java
@@ -9,7 +9,6 @@ import org.springframework.web.bind.annotation.*;
 import com.proyecto.encuentranos.modelos.ServicioPrestadorModelo;
 import com.proyecto.encuentranos.servicios.ServicioPrestadorServicio;
 
-@CrossOrigin(origins = "http://localhost:3000/")
 @RestController
 @RequestMapping("/servicio-prestador")
 public class ServicioPrestadorControlador {
