+++ registerSpringBoot/src/main/resources/templates/index.html
@@ -7,12 +7,8 @@
 <h1>Endpoints</h1>
 <h2>Private</h2>
 <ul>
-	<li><a th:href="@{/messages}">Ser cliente</a></li>
-</ul>
-
-<h2>Public</h2>
-<ul>
-	<li><a th:href="@{/messages(proveedor=true)}">Ser Proveedor</a></li>
+	<li><a th:href="@{/messages?tipoUsuarioId=1}">Ser cliente</a></li>
+	<li><a th:href="@{/messages?tipoUsuarioId=2}">Ser Proveedor</a></li>
 </ul>
 </body>
 </html>
